backend/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.8
langchain==0.1.16
elevenlabs==0.2.21
pydub==0.25.1
numpy==1.24.3
PyDrive==1.3.1
pinecone-client==2.2.4
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
PyDrive==1.3.1
wheel==0.41.2
telnyx==2.0.0
langchain-community==0.0.32
faiss-cpu==1.7.4